move UserInfo to UserCard components
