+++ src/features/users/components/UserCard/components/UserInfo/index.js
@@ -1,0 +1,2 @@
+import UserInfo from './UserInfo'
+export default UserInfo
